@@ -6,13 +6,13 @@
 
   <groupId>me.vermeil</groupId>
   <artifactId>Terminator</artifactId>
-  <version>v1.3.2</version>
+  <version>v1.4.2</version>
   <packaging>jar</packaging>
 
   <name>Terminator</name>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>21</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
 
@@ -23,8 +23,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.13.0</version>
         <configuration>
-          <source>${java.version}</source>
-          <target>${java.version}</target>
+          <source>21</source>
+          <target>21</target>
         </configuration>
       </plugin>
       <plugin>
